+++ tests/phrase_test.py
@@ -7,6 +7,162 @@ from project_toki.word import Word
 
 
 class TestPhrase(unittest.TestCase):
+    def _compare_trees(self, out_tree: Phrase, ref_tree: Phrase) -> None:
+        self.assertEqual(out_tree, ref_tree, msg=f"\n{out_tree}\n!={ref_tree}")
+
+    def test_constructors(self):
+        phrase_as_str: str = str(
+            """
+TEXT
+└── SENTENCE
+    ├── NOUN_PHRASE
+    │   ├── NOUN: "toki"
+    │   └── ADJECTIVE_PHRASE
+    │       └── ADJECTIVE: "pona"
+    ├── PARTICLE: "li"
+    ├── VERB_PHRASE
+    │   └── ADJECTIVE_PHRASE
+    │       └── ADJECTIVE: "pona"
+    ├── PARTICLE: "a"
+    └── !
+        """,
+        )
+        phrase: Phrase = Phrase(
+            "TEXT",
+            children=[
+                Phrase(
+                    "SENTENCE",
+                    children=[
+                        Phrase(
+                            "NOUN_PHRASE",
+                            children=[
+                                Phrase(Word("toki", PartOfSpeech.NOUN)),
+                                Phrase(
+                                    "ADJECTIVE_PHRASE",
+                                    children=[
+                                        Phrase(Word("pona", PartOfSpeech.ADJECTIVE)),
+                                    ],
+                                ),
+                            ],
+                        ),
+                        Phrase(Word("li", PartOfSpeech.PARTICLE)),
+                        Phrase(
+                            "VERB_PHRASE",
+                            children=[
+                                Phrase(
+                                    "ADJECTIVE_PHRASE",
+                                    children=[
+                                        Phrase(Word("pona", PartOfSpeech.ADJECTIVE)),
+                                    ],
+                                ),
+                            ],
+                        ),
+                        Phrase(Word("a", PartOfSpeech.PARTICLE)),
+                        Phrase(Punctuation("!")),
+                    ],
+                ),
+            ],
+        )
+        self.assertEqual(
+            str(phrase),
+            phrase_as_str.strip(),
+        )
+        self._compare_trees(
+            Phrase.from_str(phrase_as_str),
+            phrase,
+        )
+        self._compare_trees(
+            Phrase.from_lines(phrase_as_str.splitlines()),
+            phrase,
+        )
+
+    def test_from_lines(self):
+        self._compare_trees(
+            Phrase.from_lines(
+                [
+                    "X",
+                ],
+            ),
+            Phrase("X"),
+        )
+        self.assertRaises(
+            ValueError,
+            Phrase.from_lines,
+            [
+                "X",
+                "X",
+            ],
+        )
+        self._compare_trees(
+            Phrase.from_lines(
+                [
+                    "X",
+                    "└── X",
+                ],
+            ),
+            Phrase(
+                "X",
+                children=[
+                    Phrase("X"),
+                ],
+            ),
+        )
+        self.assertRaises(
+            ValueError,
+            Phrase.from_lines,
+            [
+                "X",
+                "├── X",
+            ],
+        )
+        self._compare_trees(
+            Phrase.from_lines(
+                [
+                    "X",
+                    "└── X",
+                    "    └── X",
+                ],
+            ),
+            Phrase(
+                "X",
+                children=[
+                    Phrase(
+                        "X",
+                        children=[
+                            Phrase("X"),
+                        ],
+                    ),
+                ],
+            ),
+        )
+        self.assertRaises(
+            ValueError,
+            Phrase.from_lines,
+            [
+                "X",
+                "├── X",
+                "│   └── X",
+            ],
+        )
+        self.assertRaises(
+            ValueError,
+            Phrase.from_lines,
+            [
+                "X",
+                "└── X",
+                "    ├── X",
+            ],
+        )
+        self.assertRaises(
+            ValueError,
+            Phrase.from_lines,
+            [
+                "X",
+                "└── X",
+                "│   └── X",
+            ],
+        )
+
     def test_diff(self):
         phrase1: Phrase = Phrase(
             "TEXT",
@@ -94,7 +250,5 @@ class TestPhrase(unittest.TestCase):
 X     │       └── ADJECTIVE: "pona"     │       └── ADJECTIVE: "ike"
 X     ├── PARTICLE: "a"                 └── !
 X     └── !
-            """[
-                1:
-            ].rstrip(),
+            """.rstrip(),
         )